Rename passkey package to webauthn
